moped_service/requirements.txt

# requirements.txt
Django==4.2.7
djangorestframework==3.14.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-decouple==3.8
ruff==0.15.1